# Core: De-duplicate babel use in core