--- a/code/.eslintignore
+++ b/code/.eslintignore
@@ -19,3 +19,4 @@ ember-output
 !.storybook
 core/assets
 core/src/core-server/utils/__search-files-tests__
+core/report