--- a/code/.prettierignore
+++ b/code/.prettierignore
@@ -1,3 +1,4 @@
 *.mdx
 
 /.nx/cache
+core/report