--- a/code/core/scripts/entries.ts
+++ b/code/core/scripts/entries.ts
@@ -37,6 +37,7 @@ export const getEntries = (cwd: string) => {
     define('src/manager/globals.ts', ['node'], true),
     define('src/preview/globals.ts', ['node'], true),
     define('src/cli/index.ts', ['node'], true),
+    define('src/babel/index.ts', ['node'], true),
     define('src/cli/bin/index.ts', ['node'], true),
   ];
 };