--- a/code/core/src/babel/index.ts
+++ b/code/core/src/babel/index.ts
@@ -6,7 +6,7 @@
 import { transformSync } from '@babel/core';
 import * as core from '@babel/core';
 // @ts-expect-error File is not yet exposed, see https://github.com/babel/babel/issues/11350#issuecomment-644118606
-import { File as BabelFileClass } from '@babel/core';
+import { File } from '@babel/core';
 import bg from '@babel/generator';
 import * as parser from '@babel/parser';
 import bt from '@babel/traverse';
@@ -19,6 +19,8 @@ import { babelParse } from './babelParse';
 const traverse = (bt.default || bt) as typeof bt;
 // @ts-expect-error (needed due to it's use of `exports.default`)
 const generate = (bg.default || bg) as typeof bg;
+
+const BabelFileClass = File as any;
 
 export {
   // main
@@ -37,6 +39,6 @@ export {
   babelParse,
 };
 
-export type { BabelFile } from '@babel/core';
+export type { BabelFile, NodePath } from '@babel/core';
 export type { GeneratorOptions } from '@babel/generator';
 export type { Options as RecastOptions } from 'recast';