--- a/code/core/src/common/utils/formatter.ts
+++ b/code/core/src/common/utils/formatter.ts
@@ -1,39 +1,6 @@
-import semver from 'semver';
-import { dedent } from 'ts-dedent';
-
-type Prettier = typeof import('prettier');
-type PrettierVersion = 3;
-
-let prettierInstance: Prettier | undefined;
-let prettierVersion: 3 | null = null;
-
-const getPrettier = async (): Promise<
-  { instance: undefined; version: null } | { instance: typeof import('prettier'); version: 3 }
-> => {
-  if (!prettierInstance) {
-    try {
-      prettierInstance = (await import('prettier')) as unknown as Prettier | undefined;
-      prettierVersion = prettierInstance?.version
-        ? (semver.major(prettierInstance.version) as PrettierVersion)
-        : null;
-
-      return {
-        version: prettierVersion,
-        instance: prettierInstance,
-      } as any;
-    } catch (err) {
-      return {
-        instance: undefined,
-        version: null,
-      };
-    }
-  }
-
-  return {
-    instance: prettierInstance,
-    version: prettierVersion,
-  } as any;
-};
+async function getPrettier() {
+  return import('prettier');
+}
 
 /**
  * Format the content of a file using prettier. If prettier is not available in the user's project,
@@ -42,49 +9,34 @@ const getPrettier = async (): Promise<
  */
 export async function formatFileContent(filePath: string, content: string): Promise<string> {
   try {
-    const prettier = await getPrettier();
+    const { resolveConfig, format } = await getPrettier();
+    const config = await resolveConfig(filePath);
 
-    switch (prettier.version) {
-      case 3:
-        const config = await prettier.instance.resolveConfig(filePath);
-
-        if (!config || Object.keys(config).length === 0) {
-          return await formatWithEditorConfig(filePath, content);
-        }
-
-        const result = await prettier.instance.format(content, {
-          ...(config as any),
-          filepath: filePath,
-        });
-
-        return result;
-      case null:
-      case undefined:
-        return await formatWithEditorConfig(filePath, content);
-      default:
-        console.warn(dedent`
-        Your prettier version ${
-          (prettier as any).version
-        } is not supported to format files which were edited by Storybook. 
-        Please raise an issue on the Storybook GitHub repository. 
-        Falling back to EditorConfig settings, if available.
-        `);
-        return await formatWithEditorConfig(filePath, content);
+    if (!config || Object.keys(config).length === 0) {
+      return await formatWithEditorConfig(filePath, content);
     }
+
+    const result = await format(content, {
+      ...(config as any),
+      filepath: filePath,
+    });
+
+    return result;
   } catch (error) {
     return content;
   }
 }
 
 async function formatWithEditorConfig(filePath: string, content: string) {
-  const prettier = await import('prettier');
-  const config = await prettier.resolveConfig(filePath, { editorconfig: true });
+  const { resolveConfig, format } = await getPrettier();
+
+  const config = await resolveConfig(filePath, { editorconfig: true });
 
   if (!config || Object.keys(config).length === 0) {
     return content;
   }
 
-  return prettier.format(content, {
+  return format(content, {
     ...(config as any),
     filepath: filePath,
   });