--- a/code/core/src/core-server/utils/save-story/duplicate-story-with-new-name.ts
+++ b/code/core/src/core-server/utils/save-story/duplicate-story-with-new-name.ts
@@ -1,19 +1,15 @@
 /* eslint-disable no-underscore-dangle */
+import { traverse, types } from '@storybook/core/babel';
+
 import type { CsfFile } from '@storybook/core/csf-tools';
 
-import bt from '@babel/traverse';
-import * as t from '@babel/types';
-
 import { SaveStoryError } from './utils';
-
-// @ts-expect-error (needed due to it's use of `exports.default`)
-const traverse = (bt.default || bt) as typeof bt;
 
 type In = ReturnType<CsfFile['parse']>;
 
 export const duplicateStoryWithNewName = (csfFile: In, storyName: string, newStoryName: string) => {
   const node = csfFile._storyExports[storyName];
-  const cloned = t.cloneNode(node) as t.VariableDeclarator;
+  const cloned = types.cloneNode(node) as types.VariableDeclarator;
 
   if (!cloned) {
     throw new SaveStoryError(`cannot clone Node`);
@@ -42,7 +38,7 @@ export const duplicateStoryWithNewName = (csfFile: In, storyName: string, newSto
   });
 
   // detect CSF2 and throw
-  if (t.isArrowFunctionExpression(cloned.init) || t.isCallExpression(cloned.init)) {
+  if (types.isArrowFunctionExpression(cloned.init) || types.isCallExpression(cloned.init)) {
     throw new SaveStoryError(`Creating a new story based on a CSF2 story is not supported`);
   }
 
@@ -50,7 +46,7 @@ export const duplicateStoryWithNewName = (csfFile: In, storyName: string, newSto
     Program(path) {
       path.pushContainer(
         'body',
-        t.exportNamedDeclaration(t.variableDeclaration('const', [cloned]))
+        types.exportNamedDeclaration(types.variableDeclaration('const', [cloned]))
       );
     },
   });