--- a/code/core/src/core-server/utils/save-story/duplicate-story-with-new-name.ts
+++ b/code/core/src/core-server/utils/save-story/duplicate-story-with-new-name.ts
@@ -1,5 +1,5 @@
 /* eslint-disable no-underscore-dangle */
-import { traverse, types } from '@storybook/core/babel';
+import { types as t, traverse } from '@storybook/core/babel';
 
 import type { CsfFile } from '@storybook/core/csf-tools';
 
@@ -9,7 +9,7 @@ type In = ReturnType<CsfFile['parse']>;
 
 export const duplicateStoryWithNewName = (csfFile: In, storyName: string, newStoryName: string) => {
   const node = csfFile._storyExports[storyName];
-  const cloned = types.cloneNode(node) as types.VariableDeclarator;
+  const cloned = t.cloneNode(node) as t.VariableDeclarator;
 
   if (!cloned) {
     throw new SaveStoryError(`cannot clone Node`);
@@ -38,7 +38,7 @@ export const duplicateStoryWithNewName = (csfFile: In, storyName: string, newSto
   });
 
   // detect CSF2 and throw
-  if (types.isArrowFunctionExpression(cloned.init) || types.isCallExpression(cloned.init)) {
+  if (t.isArrowFunctionExpression(cloned.init) || t.isCallExpression(cloned.init)) {
     throw new SaveStoryError(`Creating a new story based on a CSF2 story is not supported`);
   }
 
@@ -46,7 +46,7 @@ export const duplicateStoryWithNewName = (csfFile: In, storyName: string, newSto
     Program(path) {
       path.pushContainer(
         'body',
-        types.exportNamedDeclaration(types.variableDeclaration('const', [cloned]))
+        t.exportNamedDeclaration(t.variableDeclaration('const', [cloned]))
       );
     },
   });