--- a/code/core/src/core-server/utils/save-story/update-args-in-csf-file.ts
+++ b/code/core/src/core-server/utils/save-story/update-args-in-csf-file.ts
@@ -1,13 +1,9 @@
-import bt from '@babel/traverse';
-import * as t from '@babel/types';
+import { traverse, types } from '@storybook/core/babel';
 
 import { SaveStoryError } from './utils';
 import { valueToAST } from './valueToAST';
 
-// @ts-expect-error (needed due to it's use of `exports.default`)
-const traverse = (bt.default || bt) as typeof bt;
-
-export const updateArgsInCsfFile = async (node: t.Node, input: Record<string, any>) => {
+export const updateArgsInCsfFile = async (node: types.Node, input: Record<string, any>) => {
   let found = false;
   const args = Object.fromEntries(
     Object.entries(input).map(([k, v]) => {
@@ -16,28 +12,28 @@ export const updateArgsInCsfFile = async (node: t.Node, input: Record<string, an
   );
 
   // detect CSF2 and throw
-  if (t.isArrowFunctionExpression(node) || t.isCallExpression(node)) {
+  if (types.isArrowFunctionExpression(node) || types.isCallExpression(node)) {
     throw new SaveStoryError(`Updating a CSF2 story is not supported`);
   }
 
-  if (t.isObjectExpression(node)) {
+  if (types.isObjectExpression(node)) {
     const properties = node.properties;
     const argsProperty = properties.find((property) => {
-      if (t.isObjectProperty(property)) {
+      if (types.isObjectProperty(property)) {
         const key = property.key;
-        return t.isIdentifier(key) && key.name === 'args';
+        return types.isIdentifier(key) && key.name === 'args';
       }
       return false;
     });
 
     if (argsProperty) {
-      if (t.isObjectProperty(argsProperty)) {
+      if (types.isObjectProperty(argsProperty)) {
         const a = argsProperty.value;
-        if (t.isObjectExpression(a)) {
+        if (types.isObjectExpression(a)) {
           a.properties.forEach((p) => {
-            if (t.isObjectProperty(p)) {
+            if (types.isObjectProperty(p)) {
               const key = p.key;
-              if (t.isIdentifier(key) && key.name in args) {
+              if (types.isIdentifier(key) && key.name in args) {
                 p.value = args[key.name];
                 delete args[key.name];
               }
@@ -47,17 +43,19 @@ export const updateArgsInCsfFile = async (node: t.Node, input: Record<string, an
           const remainder = Object.entries(args);
           if (Object.keys(args).length) {
             remainder.forEach(([key, value]) => {
-              a.properties.push(t.objectProperty(t.identifier(key), value));
+              a.properties.push(types.objectProperty(types.identifier(key), value));
             });
           }
         }
       }
     } else {
       properties.unshift(
-        t.objectProperty(
-          t.identifier('args'),
-          t.objectExpression(
-            Object.entries(args).map(([key, value]) => t.objectProperty(t.identifier(key), value))
+        types.objectProperty(
+          types.identifier('args'),
+          types.objectExpression(
+            Object.entries(args).map(([key, value]) =>
+              types.objectProperty(types.identifier(key), value)
+            )
           )
         )
       );
@@ -100,7 +98,7 @@ export const updateArgsInCsfFile = async (node: t.Node, input: Record<string, an
             const remainder = Object.entries(args);
             if (Object.keys(args).length) {
               remainder.forEach(([key, value]) => {
-                a.pushContainer('properties', t.objectProperty(t.identifier(key), value));
+                a.pushContainer('properties', types.objectProperty(types.identifier(key), value));
               });
             }
           }
@@ -108,10 +106,12 @@ export const updateArgsInCsfFile = async (node: t.Node, input: Record<string, an
       } else {
         path.unshiftContainer(
           'properties',
-          t.objectProperty(
-            t.identifier('args'),
-            t.objectExpression(
-              Object.entries(args).map(([key, value]) => t.objectProperty(t.identifier(key), value))
+          types.objectProperty(
+            types.identifier('args'),
+            types.objectExpression(
+              Object.entries(args).map(([key, value]) =>
+                types.objectProperty(types.identifier(key), value)
+              )
             )
           )
         );