## What I did

I investigated the bundle meta file from esbuild for the core, and I noticed this:
![Screenshot 2024-08-26 at 12 27 43](https://github.com/user-attachments/assets/d48fc80d-55aa-4d98-84fc-81de4abfa518)

Notice the exact same blocks 3 times in 3 entrypoints?
That's babel getting embedded 3 times.

Another way to look at it:
![Screenshot 2024-08-26 at 12 28 26](https://github.com/user-attachments/assets/41a8321a-8f42-4323-a295-235221f505ff)

And I wondered, if this duplication can be dealt with by making an entry-point that exposes all these babel/transform related types and functions.

And the results are pretty good:

**before**:
![Screenshot 2024-08-26 at 12 11 45](https://github.com/user-attachments/assets/e5889a5a-641e-414f-836d-ea3cba1c9927)

**after**:
![Screenshot 2024-08-27 at 12 09 13](https://github.com/user-attachments/assets/13d83a9a-6913-4196-96fe-3e2f01c551ea)


**before**:
![Screenshot 2024-08-26 at 12 27 43](https://github.com/user-attachments/assets/02552303-536e-4cfb-8e2a-e324bce33ea1)

**after**:
![Screenshot 2024-08-27 at 12 09 18](https://github.com/user-attachments/assets/cab80332-fb7c-4f88-9c80-c73d63444747)

## Checklist for Contributors

### Testing

<!-- Please check (put an "x" inside the "[ ]") the applicable items below to communicate how to test your changes -->

#### The changes in this PR are covered in the following automated tests:
- [x] stories
- [x] unit tests
- [x] integration tests
- [x] end-to-end tests

#### Manual testing

Before this change prettier would always be used to format files when we storybook them.
(save from controls, modifying `main.ts` etc.)

Now that only happens if we can find `prettier` as a dep.
(it's an optional peerDependency of the core)

No further testing instruction here, because there should be no user impact.

### Documentation

<!-- Please check (put an "x" inside the "[ ]") the applicable items below to indicate which documentation has been updated. -->

- [ ] Add or update documentation reflecting your changes
- [ ] If you are deprecating/removing a feature, make sure to update
      [MIGRATION.MD](https://github.com/storybookjs/storybook/blob/next/MIGRATION.md)

## Checklist for Maintainers

- [ ] When this PR is ready for testing, make sure to add `ci:normal`, `ci:merged` or `ci:daily` GH label to it to run a specific set of sandboxes. The particular set of sandboxes can be found in `code/lib/cli/src/sandbox-templates.ts`
- [ ] Make sure this PR contains **one** of the labels below:
   <details>
     <summary>Available labels</summary>

     - `bug`: Internal changes that fixes incorrect behavior.
     - `maintenance`: User-facing maintenance tasks.
     - `dependencies`: Upgrading (sometimes downgrading) dependencies.
     - `build`: Internal-facing build tooling & test updates. Will not show up in release changelog.
     - `cleanup`: Minor cleanup style change. Will not show up in release changelog.
     - `documentation`: Documentation **only** changes. Will not show up in release changelog.
     - `feature request`: Introducing a new feature.
     - `BREAKING CHANGE`: Changes that break compatibility in some way with current major version.
     - `other`: Changes that don't fit in the above categories.
   
   </details>

### 🦋 Canary release

<!-- CANARY_RELEASE_SECTION -->

This PR does not have a canary release associated. You can request a canary release of this pull request by mentioning the `@storybookjs/core` team here.

_core team members can create a canary release [here](https://github.com/storybookjs/storybook/actions/workflows/canary-release-pr.yml) or locally with `gh workflow run --repo storybookjs/storybook canary-release-pr.yml --field pr=<PR_NUMBER>`_

<!-- CANARY_RELEASE_SECTION -->

<!-- BENCHMARK_SECTION -->

| name | before | after | diff | z | % |
| ---- | ------ | ----- | ---- | - | - |
| createSize |  0 B | 0 B | 0 B | - | - |
| generateSize |  76.4 MB | 76.4 MB | 0 B | 0.99 | 0% |
| initSize |  169 MB | 150 MB | -18.9 MB | **-227.17** | 🔰-12.6% |
| diffSize |  92.8 MB | 73.9 MB | -18.9 MB | **-1629.27** | 🔰-25.6% |
| buildSize |  7.46 MB | 7.46 MB | 267 B | **2.21** | 0% |
| buildSbAddonsSize |  1.62 MB | 1.62 MB | 0 B | 1.14 | 0% |
| buildSbCommonSize |  195 kB | 195 kB | 0 B | - | 0% |
| buildSbManagerSize |  2.3 MB | 2.3 MB | 0 B | - | 0% |
| buildSbPreviewSize |  352 kB | 352 kB | 0 B | **1.36** | 0% |
| buildStaticSize |  0 B | 0 B | 0 B | - | - |
| buildPrebuildSize |  4.46 MB | 4.46 MB | 0 B | 1.23 | 0% |
| buildPreviewSize |  3 MB | 3 MB | 267 B | **2.78** | 0% |
| testBuildSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbAddonsSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbCommonSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbManagerSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbPreviewSize |  0 B | 0 B | 0 B | - | - |
| testBuildStaticSize |  0 B | 0 B | 0 B | - | - |
| testBuildPrebuildSize |  0 B | 0 B | 0 B | - | - |
| testBuildPreviewSize |  0 B | 0 B | 0 B | - | - |



| name | before | after | diff | z | % |
| ---- | ------ | ----- | ---- | - | - |
| createTime |  6.7s | 21.1s | 14.4s | 0.58 | 68.2% |
| generateTime |  19s | 21.6s | 2.6s | 0.17 | 12% |
| initTime |  15.7s | 16s | 304ms | -1.04 | 1.9% |
| buildTime |  11.9s | 11.6s | -290ms | -1 | -2.5% |
| testBuildTime |  0ms | 0ms | 0ms | - | - |
| devPreviewResponsive |  7.8s | 7.1s | -777ms | -0.48 | -10.9% |
| devManagerResponsive |  4.7s | 4.5s | -124ms | -0.69 | -2.7% |
| devManagerHeaderVisible |  735ms | 844ms | 109ms | 0.43 | 12.9% |
| devManagerIndexVisible |  767ms | 885ms | 118ms | 0.43 | 13.3% |
| devStoryVisibleUncached |  1.3s | 1.1s | -240ms | -1.05 | -21.5% |
| devStoryVisible |  766ms | 884ms | 118ms | 0.39 | 13.3% |
| devAutodocsVisible |  732ms | 655ms | -77ms | -0.96 | -11.8% |
| devMDXVisible |  630ms | 714ms | 84ms | -0.08 | 11.8% |
| buildManagerHeaderVisible |  678ms | 681ms | 3ms | -0.96 | 0.4% |
| buildManagerIndexVisible |  683ms | 681ms | -2ms | -1.02 | -0.3% |
| buildStoryVisible |  717ms | 758ms | 41ms | -0.62 | 5.4% |
| buildAutodocsVisible |  690ms | 617ms | -73ms | **-1.45** | 🔰-11.8% |
| buildMDXVisible |  631ms | 614ms | -17ms | -1.04 | -2.8% |

<!-- BENCHMARK_SECTION -->

<!-- greptile_comment -->

## Greptile Summary

This PR introduces changes to de-duplicate Babel usage in the Storybook core, aiming to reduce bundle size and improve performance.

- Added new 'babel' export in @storybook/core package.json
- Created centralized Babel entry point in 'code/core/src/babel/index.ts'
- Updated imports across multiple files to use '@storybook/core/babel'
- Added bundle analysis report generation in 'code/core/scripts/prep.ts'
- Introduced new Babel-related exports in CLI package for internal use

<!-- /greptile_comment -->