@ndelangen 2024-08-26
How do you think this PR compares to yours here: https://github.com/storybookjs/storybook/pull/28571
@kasperpeulen ?